Add files via upload
I have done "NOUGHTS AND CROSSES" game ,also called as tic-tac-toe gamming using c programming.The game is to be played between two people (in this program between HUMAN and COMPUTER).
One of the player chooses ‘O’ and the other ‘X’ to mark their respective cells.The game starts with one of the players and the game ends when one of the players has one whole row/ column/ diagonal filled with his/her respective character (‘O’ or ‘X’).
If no one wins, then the game is said to be draw. In this program the moves taken by the computer and the human are chosen randomly. We use rand() function for this.The program is in not played optimally by both sides because the moves are chosen randomly. The program can be easily modified so that both players play optimally.If both the players play optimally then it is destined that you will never lose.Its quite an interesting gamming which is being played between the computer system and the human player.
If both the system and the human had secured same score it shows as the game has been drawn.